fix for test, seems to change: https://github.com/tactile/commons-scxml/commit/80dcc445cc34e909867dcb32570d6e2cf961ee22

+++ src/test/java/org/apache/commons/scxml2/env/javascript/JSEvaluatorTest.java
@@ -267,7 +267,7 @@ public class JSEvaluatorTest {
     public void testScriptFunctions() throws Exception {
         context.set("FIVE", 5);
         Assertions.assertEquals(5,context.get("FIVE"));
-        Assertions.assertEquals(120, evaluator.eval(context,FUNCTION), "Invalid function result");
+        Assertions.assertEquals(120d, (double)evaluator.eval(context,FUNCTION), "Invalid function result");
     }
